feat: replace exception pool with observable instance module
Remove ModuleExceptionPool and related exception pool classes.
Introduce ModuleObservableInstance with ObservableAgent for unified
state and exception tracking across all instances. Update DataPipeline,
HostedServices, and JobScheduler to use the new observable pattern.
Add PipelineState enum and refactor dashboard services accordingly.

+++ MoLibrary.DataChannel/Dashboard/Models/ChannelExceptionInfo.cs
@@ -18,12 +18,12 @@ public class ChannelExceptionInfo
     /// <summary>
     /// 当前异常数量
     /// </summary>
-    public int CurrentExceptions { get; set; }
-    
+    public long CurrentExceptions { get; set; }
+
     /// <summary>
     /// 总异常数量
     /// </summary>
-    public int TotalExceptions { get; set; }
+    public long TotalExceptions { get; set; }
     
     /// <summary>
     /// 异常池最大大小
